Extract helper code in utils.
